Se corrigieron los métodos insertar y eliminar del árbol B*. Había referencias de claves que no se modificaban correctamente y casos de overflow y underflow que tampoco estaban resueltos. En la clase BPlusTree sólo copipastie los métodos del árbol B*. Hay que hacer las modificaciones que correspondan.

+++ CapaIndices/Common/SetClaves.cpp
@@ -28,7 +28,7 @@ SetClaves* SetClaves::splitBPlus(unsigned short minClaves) {
 	SetClaves* conjClavesMayores = new SetClaves();
 	SetClaves::iterator iter = this->begin();
 	
-	for(unsigned i = 0; (iter != this->end()) && (i < minClaves); ++iter)
+	for(unsigned short i = 0; (iter != this->end()) && (i < minClaves); ++iter)
 		i += (*iter)->getTamanioEnDisco();
 	
 	for (SetClaves::iterator iterMayores = iter;
